created posting.jsx and call the component on Home and Mypage
