+++ frontend/.env.development
@@ -1,1 +1,1 @@
-REACT_APP_API_URL=http://localhost:5000/api/
+REACT_APP_API_URL=http://localhost:5001/api/
