+++ frontend/src/pages/Home.jsx
@@ -1,43 +1,13 @@
-import React, { useEffect } from "react";
-
-import { store, useAppDispatch, useAppSelector } from "../store/configureStore";
-import { fetchAllPosts, selectAllPosts } from "../slices/postSlice";
-import { validateUserSession } from "../slices/accountSlice";
+import React from "react";
+import Posting from "../Posting";
 
 function Home() {
-  const dispatch = useAppDispatch();
-  const posts = useAppSelector(selectAllPosts);
-  const PUBLIC = "public";
-
-  useEffect(() => {
-    dispatch(fetchAllPosts());
-    console.log("fetched posts!", store.getState().blog.posts);
-    dispatch(validateUserSession());
-  }, [dispatch]);
-
-  const reversedPosts = Object.values(
-    (Array.isArray(posts.posts) &&
-      posts.posts.filter((post) => post.postVisibility === PUBLIC).reverse()) ||
-      {}
-  );
-
   return (
     <div>
-      <h2>Posting in mypage</h2>
-      {reversedPosts.length > 0 ? (
-        reversedPosts.map((post, i) => (
-          <article key={i}>
-            <h3>title: {post.title}</h3>
-            <p>user id: {post.user}</p>
-            <h4>visibility: {post.postVisibility}</h4>
-            <p>content: {post.content}</p>
-            <p>posting time: {post.timestamp}</p>
-          </article>
-        ))
-      ) : (
-        <p>No public posts available</p>
-      )}
+      <h2>Home</h2>
+      <Posting postVisibility="public" />
     </div>
   );
 }
+
 export default Home;
